+++ demo-kingbase/kingbase-flyway/src/main/java/com/github/mengweijin/kingbase/FlywayKingbaseApplication.java
@@ -7,9 +7,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * @author mengweijin
  */
 @SpringBootApplication
-public class FlywayPostgreSQLApplication {
+public class FlywayKingbaseApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(FlywayPostgreSQLApplication.class, args);
+		SpringApplication.run(FlywayKingbaseApplication.class, args);
 	}
 }