+++ demo-kingbase/kingbase-liquibase/src/main/java/com/github/mengweijin/liquibase/KingbaseLiquibaseApplication.java
@@ -1,4 +1,4 @@
-package com.github.mengweijin.kingbase;
+package com.github.mengweijin.liquibase;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,9 +7,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  * @author mengweijin
  */
 @SpringBootApplication
-public class FlywayPostgreSQLApplication {
+public class KingbaseLiquibaseApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(FlywayPostgreSQLApplication.class, args);
+		SpringApplication.run(KingbaseLiquibaseApplication.class, args);
 	}
 }