+++ demo-postgresql/postgresql-liquibase/src/main/java/com/github/mengweijin/postgresql/LiquibasePostgreSQLApplication.java
@@ -1,4 +1,4 @@
-package com.github.mengweijin.postgresql;
+package com.github.mengweijin.kingbase;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;